// simulation-ui/IEC-104-UI/projects/iec60870-104-simulator/src/lib/api/v1/model/iecValueDto.ts
/**
 * IEC60870-5-104-simulator.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SinglePointValueDto } from './singlePointValueDto';
import { ScaledValueDto } from './scaledValueDto';
import { IntValueDto } from './intValueDto';
import { FloatValueDto } from './floatValueDto';
import { DoublePointValueDto } from './doublePointValueDto';


export interface IecValueDto { 
    numericValue?: IntValueDto;
    singlePointValue?: SinglePointValueDto;
    doublePointValue?: DoublePointValueDto;
    floatValue?: FloatValueDto;
    scaledValue?: ScaledValueDto;
}