// simulation-ui/IEC-104-UI/projects/iec60870-104-simulator/src/lib/api/v1/model/simulationMode.ts
/**
 * IEC60870-5-104-simulator.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SimulationMode = 'None' | 'Cyclic' | 'CyclicStatic' | 'Response';

export const SimulationMode = {

    None: 'None' as SimulationMode,

    Cyclic: 'Cyclic' as SimulationMode,

    CyclicStatic: 'CyclicStatic' as SimulationMode,

    Response: 'Response' as SimulationMode
};